+++ POS/POS/Modules/DeviceManagement/Constants/TransactionPortalActions.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace POS.Modules.DeviceManagement.Constants
+{
+    public class TransactionPortalActions
+    {
+        public const string COMMAND_GETALLMACHINES = "GetAllMachines";
+        public const string COMMAND_SETOFFLINE = "ShutdownMachine";
+        public const string COMMAND_SETONLINE = "StartupMachine";
+    }
+}